+++ src/index.ts
@@ -1,13 +1,13 @@
-export * from "./abs_list";
-export * from "./abs_position";
-export * from "./bunch";
-export * from "./bunch_ids";
-export * from "./lexicographic_string";
-export * from "./list";
-export * from "./order";
-export * from "./outline";
-export * from "./position";
-export * from "./text";
+export * from "./lists/abs_list";
+export * from "./lists/list";
+export * from "./lists/outline";
+export * from "./lists/text";
+export * from "./order/abs_position";
+export * from "./order/bunch";
+export * from "./order/bunch_ids";
+export * from "./order/lexicographic_string";
+export * from "./order/order";
+export * from "./order/position";
 export * from "./unordered_collections/position_char_map";
 export * from "./unordered_collections/position_map";
 export * from "./unordered_collections/position_set";
